requirements.txt
alabaster==1.0.0
allure-pytest==2.15.0
allure-python-commons==2.15.0
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
behave==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
cucumber-expressions==18.0.1
cucumber-tag-expressions==6.2.0
debugpy==1.8.15
execnet==2.1.1
Faker==37.4.2
gherkin-official==29.0.0
h11==0.16.0
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parse_type==0.6.4
pluggy==1.6.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
pytest-bdd==8.1.0
python-dotenv==1.1.1
referencing==0.36.2
requests==2.32.4
rich==14.2.0
roman-numerals-py==3.1.0
rpds-py==0.27.0
rstcheck-core==1.2.2
selenium==4.34.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stevedore==5.5.0
trio==0.30.0
trio-websocket==0.12.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
